% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticSolve.R
\name{devG}
\alias{devG}
\title{a wrapper for analyticSolve that allows it to be called from an optimizer. Takes the same arguments as analyticSolve.}
\usage{
devG(y, X, Zlist, Zlevels, weights, weightsC = weights, groupID,
  lmeVarDF, v0)
}
\arguments{
\item{y}{outcome measure.}

\item{X}{the X matrix.}

\item{Zlist, }{a list of matrixes with the Z values for each level.}

\item{Zlevels, }{the level corresponding to each matrix in Zlist.}

\item{weights}{level-1 weights}

\item{weightsC}{a list of conditional weights for each model level.}

\item{groupID}{a matrix containing the group ids for each level in the model.}

\item{weights}{a list of unconditional weights for each model level.}

\item{lmeVardf}{a dataframe containing the variances and covariance of the random effects, in the same format as returned from lme.}
}
\description{
a wrapper for analyticSolve that allows it to be called from an optimizer. Takes the same arguments as analyticSolve.
}
\keyword{internal}
