% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeSandwich}
\alias{makeSandwich}
\title{This function calculates robust standard errors using the sandwich estimator of the standard errors  
following Rabe-Hesketh & Skrondal 2006. For linear models, robust standard errors are returned from the main estimation function. 
This function is only used for post-hoc estimation for non-linear models.}
\usage{
makeSandwich(fittedModel)
}
\description{
This function calculates robust standard errors using the sandwich estimator of the standard errors  
following Rabe-Hesketh & Skrondal 2006. For linear models, robust standard errors are returned from the main estimation function. 
This function is only used for post-hoc estimation for non-linear models.
}
\keyword{internal}
